requirements.txt
requests
PyYAML
python-dotenv
newspaper3k
transformers
